+++ CHANGES.rst
@@ -2,8 +2,8 @@
 Changelog
 =========
 
-0.3dev
-------
+0.3
+---
 
 - Product listing is now batched.
   [rnix]